Frontpage almost finished, data is now being loaded in

+++ Rawdata_Porfolio_2/Entity Framework/Domain/episode.cs
@@ -11,10 +11,10 @@ namespace Rawdata_Porfolio_2.Pages.Entity_Framework.Domain
         [Key]
         public int Id { get; set; }
         public string Name { get; set; }
-        public Int64 Parent_Id { get; set; }
+        public long Parent_Id { get; set; }
         public Title Title { get; set; }
-        public int Season { get; set; }
-        public int Ep_Number { get; set; }
+        public short Season { get; set; }
+        public short Ep_Number { get; set; }
 
         public override string ToString()
         {
