+++ Rawdata_Porfolio_2/Entity Framework/Domain/search_results.cs
@@ -9,10 +9,10 @@ namespace Rawdata_Porfolio_2.Entity_Framework.Domain
     {
         public int User_Id { get; set; }
         public int Personality_ID { get; set; }
-        public Int64 Title_ID { get; set; }
+        public long Title_ID { get; set; }
         public string Character_Name { get; set; }
-        public int Year_Birth { get; set; }
-        public int Year_Death { get; set; }
+        public short Year_Birth { get; set; }
+        public short Year_Death { get; set; }
         public string Plot { get; set; }
         public string Title_Name { get; set; }
