+++ Webservice/ViewModels/PersonalityViewModel.cs
@@ -10,8 +10,8 @@ namespace Webservice.ViewModels
         public string Url { get; set; }
 
         public string Name { get; set; }
-        public int? Year_Birth { get; set; }
-        public int? Year_Death { get; set; }
+        public short? Year_Birth { get; set; }
+        public short? Year_Death { get; set; }
        // public string Profession { get; set; }
 
         /*
